server/requirements.txt
flask==2.3.0
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
pymysql==1.1.0
python-dotenv==1.0.0
